Plugin conventions overhaul
Changed the plugin convention setup to a single convention instance, configured specifically for KMP packages, as the other distinctions were (mostly) unused
