+++ buildSrc/src/main/kotlin/kmp-package.gradle.kts
@@ -1,14 +1,18 @@
 
+import com.android.build.gradle.internal.tasks.factory.dependsOn
 import org.jetbrains.kotlin.gradle.targets.js.yarn.YarnLockMismatchReport
 import org.jetbrains.kotlin.gradle.targets.js.yarn.YarnPlugin
 import org.jetbrains.kotlin.gradle.targets.js.yarn.YarnRootExtension
 
 plugins {
     kotlin("multiplatform")
+    // making them publishable & buildable for android
+    id("com.android.library")
 }
 
 repositories {
     mavenCentral()
+    google()
 }
 
 val StableSemVer = Regex("[0-9]+(?:\\.[0-9]+)+")
@@ -51,6 +55,7 @@ rootProject.plugins.withType(YarnPlugin::class.java) {
 }
 
 kotlin {
+    // target configuration
     jvm()
     js {
         var name = project.name
@@ -74,10 +79,34 @@ kotlin {
             }
         }
         nodejs()
+        browser()
+        generateTypeScriptDefinitions()
+        binaries.library()
     }
     mingwX64()
     linuxX64()
+    androidTarget {
+        publishLibraryVariants("release", "debug")
+    }
 
+    // source set configuration
+    sourceSets {
+        // https://kotlinlang.org/docs/multiplatform-hierarchy.html#manual-configuration
+        val nativeMain by creating {
+            dependsOn(commonMain.get())
+        }
+        linuxX64Main.get().dependsOn(nativeMain)
+        mingwX64Main.get().dependsOn(nativeMain)
+        // the reason for this custom hierarchy:
+        // https://slack-chats.kotlinlang.org/t/15994222/hello-hello-i-started-to-use-expected-actual-is-a-module-of-#735f0201-c023-485d-bc23-577addd2215c
+        val commonJvmMain by creating {
+            dependsOn(commonMain.get())
+        }
+        jvmMain.get().dependsOn(commonJvmMain)
+        androidMain.get().dependsOn(commonJvmMain)
+    }
+
+    // compiler configuration
     // silencing expect/actual warnings
     targets.all {
         compilations.all {
@@ -89,3 +118,25 @@ kotlin {
         }
     }
 }
+
+android {
+    val libs = versionCatalogs.named("libs")
+    compileSdk = libs.get("compileSdk").toInt()
+    namespace = "dev.tesserakt"
+    compileOptions {
+        // removeFirst was only added to java.util.List in JDK 21
+        // https://github.com/javalin/javalin/issues/2117#issuecomment-1960114620
+        sourceCompatibility = JavaVersion.VERSION_21
+        targetCompatibility = JavaVersion.VERSION_21
+    }
+}
+
+fun VersionCatalog.get(name: String): String = findVersion(name).orElseThrow().requiredVersion
+
+// required to prevent "<task1> uses output of task <task2>" errors introduced when generating JS libraries
+tasks.named("jsNodeProductionLibraryDistribution").dependsOn("jsTestTestDevelopmentExecutableCompileSync")
+tasks.named("jsNodeTest").dependsOn("jsProductionLibraryCompileSync")
+
+// as we're a package, and we have the JS source configured, we're creating an export task, that's collected by the
+// js-build meta module
+tasks.register("jsExport").dependsOn("jsProductionLibraryCompileSync")
