--- buildSrc/src/main/kotlin/package-conventions.gradle.kts
@@ -1,4 +1,0 @@
-plugins {
-    // expanding upon existing core package
-    id("core-package-conventions")
-}
