+++ n3/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    id("package-conventions")
+    id("kmp-package")
 }
 
 group = "n3"
